Add '@raftlabs/antd' package with Vitest setup and Babel config
This commit adds a new '@raftlabs/antd' package for react components, which includes scripts and configurations for Vitest-based unit tests and Babel transpilation. New dependencies were added to the project, and nx.json was adjusted to incorporate Vitest as the unit test runner. Additionally, tsconfig.base was updated with a new path alias.